+++ app/globals.css
@@ -121,6 +121,6 @@ body {
 }
 
 .lockup__link img {
-	height: 40px;
+	height: 38px;
 	padding: 8px;
 }
